+++ pyproject.toml
@@ -33,6 +33,7 @@ classifiers = [
   "Typing :: Typed"
 ]
 dependencies = [
+  "deprecated>=1.2,<1.3",
   "matplotlib>=3.8.0,<3.11",
   "numpy>=1.22.4,<2.3",
   "pandas>=2.1.4,<2.3",
@@ -120,6 +121,10 @@ disable_error_code = ["override"]
 
 [[tool.mypy.overrides]]
 module = "scipy.*"
+ignore_missing_imports = true
+
+[[tool.mypy.overrides]]
+module = "deprecated.*"
 ignore_missing_imports = true
 
 [[tool.mypy.overrides]]
